REPL

Documentation

type MyFSStMonad m a = StateT FilePath m a

Every operation is completed and exceptions handled within a context of parameter m in this type alias, and StateT handles preserving of the current filepath between iterations of interactive REPL.

class Monad m => MyFSMonad m where

General interface of what a monad needs to be successfully implementing required tasks.

Methods

mCheckAnythingExists :: FilePath -> m Bool

mCheckDirectoryExists :: FilePath -> m Bool

mCheckFileExists :: FilePath -> m Bool

mListContents :: FilePath -> m [FilePath]

mCreateDirectory :: FilePath -> m ()

mTouch :: FilePath -> m ()

mShortenPath :: FilePath -> m FilePath

mGetPermissions :: FilePath -> m Permissions

mReadFileContent :: FilePath -> m ByteString

mRemoveDirectory :: FilePath -> m ()

mRemoveFile :: FilePath -> m ()

mFindFilesHere :: FilePath -> String -> m [FilePath]

mWriteToFile :: FilePath -> String -> m ()

mGetFileSize :: FilePath -> m Integer

mFlush :: m ()

mCatch :: m a -> (SomeException -> m a) -> m a

mReadLine :: m String

mWriteLine :: String -> m ()

mWriteByteLine :: ByteString -> m ()

mWriteLineNoWrap :: String -> m ()

mWriteErrorMessage :: String -> m ()

printLines :: MyFSMonad m => [String] -> m ()

Helper function to print multiple lines. Uses mWriteLine to print single line.

processLS :: MyFSMonad m => [String] -> m ()

Prints contents of specified directory within a context of MyFSStMonad.

processCD :: MyFSMonad m => FilePath -> MyFSStMonad m ()

The only processor that does not have generalCatch wrapped around it and working directly in MyFSStMonad, because it is the only processor that changes state.

generalCatch :: MyFSMonad m => m () -> m ()

Wrapper for every processor operation except processCD. Catches an exception and prints it within a context of MyFSStMonad.

processCat :: MyFSMonad m => [String] -> m ()

Prints file contents within a context of MyFSStMonad

processTouch :: MyFSMonad m => [String] -> m ()

Creates an empty file, if a path before the final directory is valid and all parent directories present.

processMkdir :: MyFSMonad m => [String] -> m ()

Creates empty directory, if a path before the final directory is valid and all parent directories present.

processRemoveDirectory :: MyFSMonad m => [String] -> m ()

Removes directory recursively.

processRemoveFile :: MyFSMonad m => [String] -> m ()

Removes file.

recursiveSearch :: MyFSMonad m => String -> String -> m [String]

Helper function for processLocate, the step of the recursion.

processLocate :: MyFSMonad m => [String] -> m ()

Searches recursively for the supplied file within current directory, printing full filepaths into the context of MyFSStMonad.

processEcho :: MyFSMonad m => [String] -> m ()

Prints supplied text to the supplied file. Overwrites existing content.

formatCoolBytes :: Integer -> String

Helper function, turning a number of bytes into representation through bytes, kbytes or mbytes.

goCalcSizeAmount :: MyFSMonad m => String -> m (Integer, Integer)

Calculates the size of the directory, recursively traversing its children.

processDirinfo :: MyFSMonad m => [String] -> m ()

Prints info about the directory within the MyFSStMonad context on invocation.

processFileinfo :: MyFSMonad m => [String] -> m ()

Prints info about the file within the MyFSStMonad context on invocation.

processHelp :: MyFSMonad m => m ()

Prints help message within the MyFSStMonad context on invocation.

curGetTargetAndLiftWrapper :: MyFSMonad m => ([String] -> m ()) -> FilePath -> MyFSStMonad m ()

General purpose wrapper around a processor of a command, lifts to the context of the MyFSStMonad and wraps into an exception catcher generalCatch. Used to supply one argument to the processor.

curGetTargetAndLiftWrapperWithCur :: MyFSMonad m => ([String] -> m ()) -> FilePath -> MyFSStMonad m ()

General purpose wrapper around a processor of a command, lifts to the context of the MyFSStMonad and wraps into an exception catcher generalCatch. Used to supply two arguments to the processor.

listProcessors :: MyFSMonad m => Map String (FilePath -> MyFSStMonad m ())

A map that stores functions that process commands from the recognizable list. Commands being:

ls - processLS

help - processHelp

mkdir - processMkdir

touch - processTouch

cat - processCat

echo - processEcho

rmdirectory - processRemoveDirectory

rmfile - processRemoveFile

locate - processLocate

fileinfo - processFileinfo

dirinfo - processDirinfo

singleEntry :: IO ()

Entrypoint to the application when built with `stack build`.

Uses current user directory as a starting position in the tree.

loop :: MyFSMonad m => MyFSStMonad m ()

The main REPL function, user input processing happens there