REPLSpec

Documentation

data MyEx

Custom exception type, describing every expected incorrect situation that can occur while running tests.

Constructors

LsOnMissingDir

PermissionsOnNonExistentEntity

FileSizeOnNonExistentEntity

AttemptToReadNonExistentFile

AttemptToWriteToNonexistentFile

DeleteNonExistentDir

DeleteNonExistentFile

CreateDirOnNonExistentPath

CreateFileOnNonExistentPath

DirectoryExists

FileExists

newtype DirData x

Wrapper around directory entity, encapsulating map entity that stores children of this directory.

Constructors

DirData

Fields

_lst :: Map String x

lst :: forall x x. Iso (DirData x) (DirData x) (Map String x) (Map String x)

data FileParams

Wrapper around file entity, encapsulating file properties - name and content.

Constructors

FileParams

Fields

_filename :: String

_content :: ByteString

filename :: Lens' FileParams String

content :: Lens' FileParams ByteString

data MyFS

Node entity for the toy filesystem, can be either a directory or a file.

Constructors

Dir (DirData MyFS)

File FileParams

_File :: Prism' MyFS FileParams

_Dir :: Prism' MyFS (DirData MyFS)

data MyFSPlus

Toy filesystem entity, contains directory structure and three buffers to check connection to the inputoutputexceptions data streams.

Constructors

MyFSPlus

Fields

_tree :: MyFS

_exc :: [String]

_logs :: [String]

_input :: [String]

tree :: Lens' MyFSPlus MyFS

logs :: Lens' MyFSPlus [String]

input :: Lens' MyFSPlus [String]

exc :: Lens' MyFSPlus [String]

type TestFSMonad = ExceptT MyEx (State MyFSPlus)

Type alias for the pure calculation that is running in tests. State stores filesystem state, and ExceptT wraps a potential custom exception around this calculation.

splitChopHome :: FilePath -> [FilePath]

Helper function, splits absolute filepath using splitDirectories and chops the root / element.

generateWalkLens :: [String] -> Traversal' MyFSPlus MyFS -> Traversal' MyFSPlus MyFS

Helper function, generates Lens to the given filesystem node given absolute filepath.

accessor :: String -> Traversal' MyFSPlus MyFS -> Traversal' MyFSPlus MyFS

Helper function, wrapper around path preprocessing + Lens generation

tryWalkTree :: String -> TestFSMonad (Maybe MyFS)

Helper function, generates Lens to the given filesystem node and retrieves its value, given absolute filepath.

checkCreationPreconditions :: FilePath -> MyEx -> MyEx -> TestFSMonad ([FilePath], FilePath)

Helper function, checks following conditions and throws corresponding exceptions, if:

parent directory for directory creation does not exist - CreateDirOnNonExistentPath

directory already exists - DirectoryExists

parent directory for file creation does not exist - CreateFileOnNonExistentPath

file already exists - FileExists

checkDeletionPreconditions :: FilePath -> MyEx -> TestFSMonad ([FilePath], FilePath)

Helper function, throws DeleteNonExistentDir or DeleteNonExistentFile if attempted to delete filesystem entity that does not exist.

hRemoveDD :: [FilePath] -> [FilePath] -> FilePath

Helper function that removes `..` from supplied path.

mapElemAccessor :: Traversal' MyFSPlus MyFS -> String -> Traversal' MyFSPlus (Maybe MyFS)

Helper accessor that gives Traversal on Map element (pointer to the next node in filesystem tree).

simpleFile :: MyFS

Test data, simple file.

simpleDir :: MyFS

Test data, simple directory containing one file.

abcData :: MyFS

Test data, bottomless recursive directory structure.

dirWithDirs :: MyFS

Test data, simple folder with 3 simpleDir in it.

nonRecursive2Lvl :: MyFS

Test data, folder with 3 dirWithDirs in it.

testFiniteTree :: [String] -> MyFSPlus

Test data, wrapper around bottomless recursive directory structure.

testRecursiveTree :: [String] -> MyFSPlus

Test data, wrapper around bottomless recursive directory structure.

runCheckers :: FilePath -> MyFSPlus -> [MyFSPlus -> Expectation] -> (Either MyEx FilePath -> Expectation) -> IO ()

Wrapper for REPL execution, running the calculation and then verifying output with supplied checkers.

noSuchCommand :: String

Text message printed if unknown command printed.

noSuchDirectory :: String

Text message printed if processCD invoked with invalid filesystem arguments.

pr :: String

Shortcut for prompt string.

simpleDirinfoList :: String -> [String]

Test data, expected result of dirinfo command

simpleFileList :: String -> [String]

Test data, expected result of fileinfo command

simpleLocateList :: [String]

Test data, expected result of locate command on nonRecursive2Lvl

mainLibSpec :: IO ()

Entry point for REPL tests